# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Break Buddy API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: number;
  name: string;
  /** @nullable */
  teamId?: number | null;
  /** @nullable */
  teamName?: string | null;
  createdAt: string;
}

export interface UserInput {
  /**
     * @minLength 1
     * @maxLength 50
     */
  name: string;
}

export interface UserStats {
  userId: number;
  todayBreaks: number;
  weeklyBreaks: number;
  todayHydration: number;
  weeklyHydration: number;
}

export interface Team {
  id: number;
  name: string;
  inviteCode: string;
  createdAt: string;
  memberCount: number;
}

export interface TeamInput {
  /**
     * @minLength 1
     * @maxLength 50
     */
  name: string;
  userId: number;
}

export interface TeamJoinInput {
  inviteCode: string;
  userId: number;
}

export interface LeaderboardMember {
  userId: number;
  name: string;
  weeklyBreaks: number;
  todayBreaks: number;
  weeklyHydration: number;
  rank: number;
  /** @nullable */
  medal?: string | null;
}

export interface Leaderboard {
  teamId: number;
  teamName: string;
  weekStart: string;
  members: LeaderboardMember[];
}

export interface BreakEntry {
  id: number;
  userId: number;
  breakType: string;
  completedAt: string;
}

export interface BreakInput {
  userId: number;
  breakType: string;
}