# lib/api-zod/src/generated/types/breakEntry.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Break Buddy API
 * OpenAPI spec version: 0.1.0
 */

export interface BreakEntry {
  id: number;
  userId: number;
  breakType: string;
  completedAt: string;
}