# lib/api-zod/src/generated/types/leaderboard.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Break Buddy API
 * OpenAPI spec version: 0.1.0
 */
import type { LeaderboardMember } from './leaderboardMember';

export interface Leaderboard {
  teamId: number;
  teamName: string;
  weekStart: string;
  members: LeaderboardMember[];
}