# lib/api-zod/src/generated/types/leaderboardMember.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Break Buddy API
 * OpenAPI spec version: 0.1.0
 */

export interface LeaderboardMember {
  userId: number;
  name: string;
  weeklyBreaks: number;
  todayBreaks: number;
  weeklyHydration: number;
  rank: number;
  /** @nullable */
  medal?: string | null;
}