# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Break Buddy API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  name: string;
  /** @nullable */
  teamId?: number | null;
  /** @nullable */
  teamName?: string | null;
  createdAt: string;
}