# lib/api-zod/src/generated/types/userInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Break Buddy API
 * OpenAPI spec version: 0.1.0
 */

export interface UserInput {
  /**
     * @minLength 1
     * @maxLength 50
     */
  name: string;
}