# lib/api-zod/src/generated/types/userStats.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Break Buddy API
 * OpenAPI spec version: 0.1.0
 */

export interface UserStats {
  userId: number;
  todayBreaks: number;
  weeklyBreaks: number;
  todayHydration: number;
  weeklyHydration: number;
}